requirements.txt
Django
djangorestframework
pandas 
numpy 
scikit-learn